Simplify loading of Segment script
