+++ common/server-data/__mocks__/index.ts
@@ -17,11 +17,11 @@ export function clear(): void {
 export async function getServerData(): Promise<ServerData> {
   return {
     toggles: {},
-    hasAnalyticsConsent: false,
     prismic: {
       globalAlert: emptyGlobalAlert(),
       popupDialog: emptyPopupDialog(),
       collectionVenues: emptyPrismicQuery<CollectionVenuePrismicDocument>(),
     },
+    hasAnalyticsConsent: false,
   };
 }
